requirements.txt
pandas==1.3.4
ta==0.8.0
requests==2.26.0
pyyaml==6.0
sphinx==4.3.1
karma-sphinx-theme
